+++ test/fixtures/action-good/action.js
@@ -1,0 +1,3 @@
+module.exports = async ({ console }) => {
+	console.log('Hi from the action!');
+};