+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.9.0
+
+ * feat(extension): Added support actions.
+ * fix(terminal): One time keypress listener was not removing sigint handler after firing causing
+   programs to appear hung.
+ * chore: Updated dependencies.
+
 # 1.8.9 (Nov 15, 2020)
 
  * fix(cli): Fixed bug where banner was rendered twice if `autoHideBanner` is enabled and an the